swagger solution
