+++ src/main/java/org/cnu/realcoding/myfirstspringbootapp6/Service/DogManagementService.java
@@ -1,8 +1,8 @@
-package org.cnu.realcoding.service;
+package org.cnu.realcoding.myfirstspringbootapp6.Service;
 
 import lombok.Getter;
-import org.cnu.realcoding.domain.Dog;
-import org.cnu.realcoding.exception.DogNotFoundException;
+import org.cnu.realcoding.myfirstspringbootapp6.domain.Dog;
+import org.cnu.realcoding.myfirstspringbootapp6.exception.DogNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -10,21 +10,21 @@ import java.util.List;
 
 @Service
 public class DogManagementService {
+
     @Getter
-    private List<Dog> dogs = new ArrayList<>();
+    private List<Dog> dogs=new ArrayList<>();
 
     public void insertDog(Dog dog) {
         dogs.add(dog);
     }
 
-    public Dog getDogByName(String name) {
 
+    public Dog getDogByname(String name) {
         for (Dog dog : dogs) {
             if (dog.getName().equals(name)) {
                 return dog;
             }
         }
-
         throw new DogNotFoundException();
     }
 }
