+++ src/main/java/org/cnu/realcoding/myfirstspringbootapp6/exception/DogNotFoundException.java
@@ -1,4 +1,4 @@
-package org.cnu.realcoding.exception;
+package org.cnu.realcoding.myfirstspringbootapp6.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
